+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Other blocks are:
 - **BLS**: Ballast failed to parse the `Layout` block if present.
 - **BLS**: Ballast failed to read all parameters in `Layout` block, which could result in the curves not being displayed.
 - **BLS**: `--gen-ini` and `--gen-longini` do not print anymore an extra `\` at the start of the file.
+- **LIB**: Fixed parsing of CSV files in `Spectrum` class.
 
 ### Changed
 - **API**: `QData.get()` does not raise an AttributeError if the the quantity is the field is not found, but simply `None`.